Extract BaseSDXLRunner
