+++ tt-media-server/domain/base_image_request.py
@@ -4,7 +4,7 @@
 
 from domain.base_request import BaseRequest
 from pydantic import Field
-from typing import Optional
+from typing import List, Optional, Tuple, Union
 
 
 class BaseImageRequest(BaseRequest):
@@ -14,6 +14,12 @@ class BaseImageRequest(BaseRequest):
     seed: Optional[int] = None
     guidance_scale: float = Field(..., ge=1.0, le=20.0)
     number_of_images: Optional[int] = Field(default=1, ge=1, le=4)
+    prompt_2: Optional[str] = None
+    negative_prompt_2: Optional[str] = None
+    crop_coords_top_left: Optional[Tuple[int, float]] = Field(default=(0, 0))
+    guidance_rescale: Optional[float] = Field(default=0.0, ge=0.0, le=1.0)
+    timesteps: Optional[List[Union[int, float]]] = None
+    sigmas: Optional[List[Union[int, float]]] = None
 
     def update_object(self, **kwargs):
         for key, value in kwargs.items():
