--- tt-media-server/domain/image_generate_request.py
@@ -1,17 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-#
-# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
-
-from typing import Optional
-
-from domain.base_image_request import BaseImageRequest
-from pydantic import Field
-from typing import Optional, Union, List, Tuple
-
-class ImageGenerateRequest(BaseImageRequest):
-    prompt_2: Optional[str] = None
-    negative_prompt_2: Optional[str] = None
-    crop_coords_top_left: Optional[Tuple[int, float]] = Field(default=(0, 0))
-    guidance_rescale: Optional[float] = Field(default=0.0, ge=0.0, le=1.0)
-    timesteps: Optional[List[Union[int, float]]] = None
-    sigmas: Optional[List[Union[int, float]]] = None
