+++ tt-media-server/domain/image_to_image_request.py
@@ -2,13 +2,11 @@
 #
 # SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
 
-from typing import Optional
-
 from domain.base_image_request import BaseImageRequest
 from typing import Optional
 
 class ImageToImageRequest(BaseImageRequest):
-    image: str  # Base64-encoded image
+    image: str  
     strength: Optional[float] = 0.3
     aesthetic_score: Optional[float] = 6.0
     negative_aesthetic_score: Optional[float] = 2.5
